refactoring for better performance. yanking bittensor metagraph stuff from main streamlit
